Move News widget to Overiew module and let it inherit from WidgetComponent
