+++ modules/grids/app/views/grids/widgets/news/show.html.erb
@@ -27,4 +27,4 @@ See COPYRIGHT and LICENSE files for more details.
 
 ++#%>
 
-<%= render Grids::Widgets::NewsComponent.new(project: @project, current_user:) %>
+<%= render Overviews::Widgets::NewsComponent.new(@project) %>
