+++ modules/overviews/app/components/overviews/overview_grid_component.html.erb
@@ -30,7 +30,10 @@ See COPYRIGHT and LICENSE files for more details.
 <%=
   render(Grids::WidgetGridComponent.new) do |grid|
     grid.with_widget(Overviews::Widgets::DescriptionComponent, @project)
-    grid.with_widget(Overviews::Widgets::SubitemsComponent, @project)
     grid.with_widget(Overviews::Widgets::ProjectStatusComponent, @project)
+    grid.with_widget(Overviews::Widgets::SubitemsComponent, @project)
+    if OpenProject::FeatureDecisions.beta_widgets_active?
+      grid.with_widget(Overviews::Widgets::NewsComponent, @project)
+    end
   end
 %>
